Add CI GitHub Actions workflow
This continuous integration workflow cross-compiles Nickel to WASM,
then builds the rest of the website, and deploys the whole thing to
Netlify using the Netlify CLI. Netlify automatic builds should
therefore be disabled. We use GitHub Actions for the build instead of
Netlify because that way we can use Nix.

The Nickel repo is tracked as a submodule. Changes to the Nickel repo
are *not* automatically reflected on the website. Instead, the website
tracks releases (once these exist). Deploying changes to Nickel on the
website requires updating the submodule reference by hand.