Fill in missing website metadata for SEO
This commit removes the embarrassing "Sample" fillers in the website's
metadata, so that thumbnails and the like will show proper descriptions.

The word "Write" is also traded for "Manage", because "Write complex
configuration" sounds like the goal is to write complex configuration,
but that we could write simple ones as well. I feel like "Manage" better
reflects the fact that the complexity is there, and we have to handle it
one way or another. And Nickel is a good way.